requirements.txt
acres==0.3.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.13.3
bm25s==0.2.8
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
comm==0.2.2
configobj==5.0.9
configparser==7.1.0
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
etelemetry==0.3.1
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.10.0
fastapi==0.115.11
filelock==3.17.0
filetype==1.2.0
fitz==0.0.1.dev2
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.34.0
isodate==0.6.1
jedi==0.19.2
jiter==0.8.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
llama-cloud==0.1.14
llama-cloud-services==0.6.5
llama-index==0.12.23
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.23.post2
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.8
llama-index-llms-openai==0.3.25
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.6
llama-index-readers-llama-parse==0.4.0
llama-index-retrievers-bm25==0.5.2
llama-index-vector-stores-faiss==0.3.0
llama-parse==0.6.4.post1
looseversion==1.3.0
lxml==5.3.1
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
nltk==3.9.1
numpy==2.2.3
openai==1.65.4
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pdfminer.six==20231228
pdfplumber==0.11.5
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.3.0
prov==2.0.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.28
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.4
Pygments==2.19.1
pyparsing==3.2.1
pypdf==5.3.1
pypdfium2==4.30.1
PyStemmer==2.2.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==26.2.1
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
scipy==1.15.2
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.46.1
striprtf==0.0.26
tenacity==9.0.0
tiktoken==0.9.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
wrapt==1.17.2
yarl==1.18.3
